Change int to float

+++ Calculator.py
@@ -1,7 +1,7 @@
 print("Welcome to my Calculator")
 
-num1 = int(input("Enter the first number: "))
-num2 = int(input("Enter the second number: "))
+num1 = float(input("Enter the first number: "))
+num2 = float(input("Enter the second number: "))
 
 method = input("Enter operation name (+, -, *, /): ")
